pin helm image to 3.3.1
I haven't validated this yet, but the `make helm-diff` step in the deploy job
started breaking today just as the 3.3.2 tag of alpine/helm was shipped.
Tracking the base image for an oci-build-task build in a resource seems to be
a bit tricky, so I'll commit this "pinning" operation in the dockerfile itself
to see if it resolves the issue.

Signed-off-by: Jamie Klassen <jklassen@vmware.com>